api-src/requirements.txt
Flask==2.0.3
nsj_gcf_utils==1.0.1
pydantic==1.9.0
google-cloud-secret-manager==2.0.0
